refactor: move chart templates for reuse
