+++ TrinityFrontend/src/templates/charts/RechartsChartRenderer.tsx
