+++ TrinityFrontend/src/templates/charts/chart.css
